% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animals.df}
\alias{animals.df}
\title{Animals dataset}
\format{
A data frame with 50 elements and 2 columns:
\describe{
\item{animal.size}{a character string representing the animal size in inches (i.e "12in")}
\item{species}{a character string containing a specific species name (i.e "Panther")}
}
}
\usage{
data(animals.df)
}
\description{
This randomly generated dataset pairs animal sizes with species names drawn from 'animals.dict'.
Since elements in the dataset represent specific species rather than broad categories, this dataset can demonstrate how 'str_clean' could be used to group values into higher-level classifications.
}
\keyword{datasets}
