% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animals.dict}
\alias{animals.dict}
\title{Animals dictionary}
\format{
A named list with 5 elements, each element corresponding to a character vector of species names belonging to a category.
\describe{
\item{feline}{house cat, tiger, panther, lion, jaguar}
\item{canine}{wolf, fox, coyote, house dog, jackal}
\item{birds}{sparrow, pigeon, chicken, falcon, eagle}
\item{turtles}{snapping turtle, box turtle, tortoise}
\item{insects}{beetle, ants, wasp, bees, mantis}
}
}
\usage{
data(animals.dict)
}
\description{
A dictionary mapping broad animal categories to select member species.
Like invest.dict, this corresponds with animals.df and used for the 'dict' argument in 'str_clean()', resulting in the recoding of animal species into their broader category (i.e "tiger" -> "feline")
}
\keyword{datasets}
