% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_width.R
\name{fixed_width}
\alias{fixed_width}
\alias{fixed_width.character}
\alias{fixed_width.numeric}
\alias{fixed_width.integer}
\alias{fixed_width.default}
\title{To format numbers and strings. See help("format") for more information.}
\usage{
fixed_width(x, width, ...)

\method{fixed_width}{character}(x, width, justify = "left", ...)

\method{fixed_width}{numeric}(x, width, digits = 0, ...)

\method{fixed_width}{integer}(x, width, digits = 0, ...)

\method{fixed_width}{default}(x, width, digits = 0, ...)
}
\description{
To format numbers and strings. See help("format") for more information.
}
