% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{format_each}
\alias{format_each}
\title{Format numbers, rounding to a given number of digits}
\usage{
format_each(input_data, ..., digits)
}
\arguments{
\item{input_data}{tbl containing columns to total}

\item{\dots}{additional parameters passed on to methods}

\item{digits}{number of digits}
}
\description{
Format numbers, rounding to a given number of digits
}
