% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mortality_rate.R
\name{format_mortality_rate}
\alias{format_mortality_rate}
\title{format_mortality_rate}
\usage{
format_mortality_rate(
  x,
  units = "death/Mperson",
  digits = NULL,
  signif = Inf,
  suffix = NULL
)
}
\arguments{
\item{x}{value}

\item{units}{to express \code{value} in}

\item{digits}{passed to \code{\link[=format_count]{format_count()}}}

\item{signif}{passed to \code{\link[base:Round]{base::signif()}}}

\item{suffix}{(optional) will default to "/M" if units are "/Mperson"}
}
\description{
format_mortality_rate
}
