% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_percent_change.R
\name{format_percent_change}
\alias{format_percent_change}
\title{format_percent_change}
\usage{
format_percent_change(x, digits = 0, na = "", zero = "0\%")
}
\arguments{
\item{x}{numeric}

\item{na}{value to use if \code{x} is \code{NA}; defaults to an empty string}

\item{zero}{value to use if \code{x} is unity; defaults to "0%" (unsigned, no digits).}
}
\description{
Format a relative change.
}
