% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_percentage.R
\name{format_percentage}
\alias{format_percentage}
\title{Format percentages}
\usage{
format_percentage(x, digits = 0, ...)
}
\arguments{
\item{x}{(numeric)}
}
\value{
(character)
}
\description{
Express fractions as percentages, to the precision given by \code{digits}.
}
\examples{
format_percentage(1.00)
format_percentage(2/3)
format_percentage(2/3, digits = 3)

}
