% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_range.R
\name{format_range}
\alias{format_range}
\title{format_range}
\usage{
format_range(x, digits = NULL, sep = " to ", na.rm = TRUE)
}
\arguments{
\item{x}{(numeric) vector}

\item{digits}{(numeric) passed to \code{\link[=format_digits]{format_digits()}}}

\item{sep}{(character) separator, like " to " or "-"}

\item{na.rm}{(logical)}
}
\description{
format_range
}
\examples{
format_range(1:10)
}
