% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{format_scientific}
\alias{format_scientific}
\title{Scientific format (Unicode)}
\usage{
format_scientific(x, ...)
}
\description{
Scientific format (Unicode)
}
\examples{
\dontrun{
format_scientific(1.23e-5, digits = 3)
format_scientific(11, digits = 2)
format_scientific(9.0, digits = 2)
}
}
