% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCommas.R
\name{formatCommas}
\alias{formatCommas}
\title{Format table columns}
\usage{
formatCommas(table, columns)
}
\description{
Format table columns
}
\note{
For use with \code{DT} package
}
\seealso{
formatID
}
