% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatID.R
\name{formatID}
\alias{formatID}
\title{formatID}
\usage{
formatID(table, columns, prefix = "#")
}
\description{
formatID
}
\note{
For use with \code{DT} package

Exploits the (private!) \code{DT:::formatColumns} function
}
\seealso{
formatCommas
}
