% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_TeX.R
\name{glue_TeX}
\alias{glue_TeX}
\title{glue_TeX}
\usage{
glue_TeX(..., output = c("expression", "text", "ast"))
}
\arguments{
\item{...}{passed to \code{\link[glue:glue]{glue::glue()}}}

\item{output}{passed to \code{\link[latex2exp:TeX]{latex2exp::TeX()}}}
}
\value{
suitable for use with ggplot, etc.
}
\description{
glue_TeX
}
