% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_digits.R
\name{guess_digits}
\alias{guess_digits}
\title{guess_digits}
\usage{
guess_digits(x, max = base::max(c(6, getOption("digits"))))
}
\description{
guess_digits
}
\examples{
guess_digits(c(1, 0.6, 0.25))
guess_digits(0.66)
guess_digits(0.667)

}
\seealso{
\itemize{
\item \code{\link[=format_digits]{format_digits()}}
\item \code{\link[=format_SI]{format_SI()}}
\item \code{\link[=format_qty]{format_qty()}}
}
}
