% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invest.df}
\alias{invest.df}
\title{Investment co-investors dataset:}
\format{
A data frame with 5 rows and 2 columns:
\describe{
\item{investment}{Numeric investment ID}
\item{co.investors}{A semicolon-separated character string of noisy co-investor names.}
}
}
\usage{
data(invest.df)
}
\description{
This dataset contains fake investment entries with intentionally noisy co-investor string values (in "co.investors") to demonstrate str_clean()'s capabilities.
The raw string data includes inconsistent casing, typos, alternative names and punctuation variations.
The dataset was derived from invest.dataset.csv made in Excel by me.
}
\keyword{datasets}
