% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invest.dict}
\alias{invest.dict}
\title{Investment co-investor dictionary}
\format{
A data frame with 5 elements. Each element is a character vector of lowercase synonyms corresponding to a firm name known as the "key"
\describe{
\item{KKR}{kohlberg kravis roberts, kkr}
\item{Chase}{chase, jp morgan, jp morgan chase}
\item{BofA}{bofa, bofa securities, inc., bank of america}
\item{Blackstone}{bx, blackstone, blackstone inc.}
\item{UBS}{ubs, ubs group ag}
}
}
\usage{
data(invest.dict)
}
\description{
This dictionary maps firm names to common abbreviations, subsidiaries and mis-spellings, etc.
It is complimentary to invest.df and used in str_clean()
}
\keyword{datasets}
