% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{letters.df}
\alias{letters.df}
\title{Random character dataset}
\format{
A dataframe with 50 rows and 2 columns:
\describe{
\item{random.char}{a single uppercase letter sampled randomly from LETTERS}
\item{mixed_random.char}{randomly lowercase converted letters in 'random.char', producing a vector of upper and lowercase values. }
}
}
\usage{
data(letters.df)
}
\description{
A randomly generated dataset containing uppercase and mixed-case single letters.
This dataset is intended to demonstrate the parameter 'case.sensitive' argument in 'str_unique()' and 'str_find'. where letter casing affects how unique values are counted and identified.
}
\keyword{datasets}
