% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_integers.R
\name{pack_integers}
\alias{pack_integers}
\title{Packed representation of integers (R notation).}
\usage{
pack_integers(x)
}
\arguments{
\item{x}{vector of integers (integer-ish is OK)}
}
\value{
string
}
\description{
Packed representation of integers (R notation).
}
\examples{
x <- c(1, 3, 4:7, 9)
pack_integers(x)
}
