% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_integers.R
\name{parse_integers}
\alias{parse_integers}
\title{Parse a single character string representing a vector of integers}
\usage{
parse_integers(x)
}
\description{
Parse a single character string representing a vector of integers
}
\note{
TODO: ensure no rounding occurs (validate input; no decimal places)?
}
\examples{
parse_integers("1")
parse_integers("c(2, 3)")
parse_integers("c(5:9)")
parse_integers("5:9")
parse_integers("NA")
parse_integers("foo")

}
\seealso{
unpack_ids
}
