% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_percentage.R
\name{parse_percentage}
\alias{parse_percentage}
\title{Parse percentages, dividing by 100}
\usage{
parse_percentage(x, na = c("", "NA", "None"), ...)
}
\arguments{
\item{x}{character}

\item{na}{character}

\item{\dots}{Further arguments to \code{readr::parse_double}.}
}
\value{
numeric
}
\description{
Converts strings into real numbers. Divides by 100.
}
\note{
Throws an error if values are out of range. Future versions may offer custom handling.
}
\examples{
parse_percentage("12.0\%")

}
