% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision}
\alias{precision}
\title{precision}
\usage{
precision(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a power of ten (10, 100, 1000, etc.)
}
\description{
precision
}
\note{
Lifted from \pkg{scales} (doesn't export \code{precision})
}
