% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename.default}
\alias{rename.default}
\title{rename.character}
\usage{
\method{rename}{default}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{further arguments, as with \code{\link[dplyr:rename]{dplyr::rename()}}}
}
\description{
rename.character
}
