% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_blank.R
\name{replace_blank}
\alias{replace_blank}
\title{Replace blank character strings with a default value}
\usage{
replace_blank(x, value = NA)
}
\arguments{
\item{x}{vector containing ""s}

\item{value}{replacement value}
}
\description{
Replace blank character strings with a default value
}
\examples{
x <- c("foo", "", "bar")
replace_blank(x)
replace_blank(x, value = "-8888")

}
