% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_zero.R
\name{replace_zero}
\alias{replace_zero}
\title{Replace zeros with a default value}
\usage{
replace_zero(x, value = NA)
}
\arguments{
\item{x}{vector containing zeros}

\item{value}{replacement value}
}
\description{
Replace zeros with a default value
}
\examples{
x <- c(0, 1, 2, 3)
replace_zero(x)
replace_zero(x, value = NaN)

}
