% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_and.R
\name{str_and}
\alias{str_and}
\title{str_and}
\usage{
str_and(x, ..., .sep = ",")
}
\description{
A more human-friendly version of \code{str_csv()}.
}
\examples{
x <- c("apples", "oranges", "pears")
str_and(x)

}
\seealso{
\code{\link[=str_or]{str_or()}}
}
