% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_begins.R
\name{str_begins}
\alias{str_begins}
\title{Detect strings beginning with some substring}
\usage{
str_begins(x, pattern)
}
\arguments{
\item{x}{character vector}

\item{pattern}{substring that you want to detect}
}
\description{
helpful to combine with \code{filter()}
}
\examples{
x <- c("C22240098", "C22250098", "C22240189", "G1003260")
x \%>\% str_begins("G")
x \%>\% str_begins("C22")

library(tidyverse)
df <- data_frame(src_code = x)
df \%>\% filter(src_code \%>\% str_begins("G"))
df \%>\% filter(src_code \%>\% str_begins("C22"))

}
