% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_cross.R
\name{str_cross}
\alias{str_cross}
\title{str_cross}
\usage{
str_cross(..., sep = "")
}
\description{
Full crossing (all combinations) of given strings
}
\examples{
str_cross(c("foo", "bar"), c("baz", "bap"), sep = "-")

}
