% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_csv.R
\name{str_csv}
\alias{str_csv}
\title{Paste values together, separated by commas}
\usage{
str_csv(..., collapse = ", ")
}
\arguments{
\item{\dots}{character}
}
\description{
Paste values together, separated by commas
}
\examples{
str_csv(1:3)
str_csv(1:3, "A", c("B", "C"))

}
