% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_date.R, R/str_datestamp.R
\name{str_date}
\alias{str_date}
\alias{str_datestamp}
\title{str_date}
\usage{
str_date(format = "\%Y-\%m-\%d", date = Sys.Date(), ...)

str_datestamp(format = "\%Y\%m\%d", date = Sys.Date(), ...)
}
\arguments{
\item{format}{see \code{\link[=format]{format()}}}

\item{date}{defaults to \code{\link[=Sys.Date]{Sys.Date()}}.}

\item{...}{passed to \code{format()}.}
}
\description{
Output a datestamp.
}
\details{
\code{str_date()} defaults to \code{YYYY-mm-dd} (ISO 9601), while \code{str_datestamp()} defaults to a more concise \code{YYYYmmdd}. You can override either one.
}
