% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_datestamp.R
\name{str_datestamp}
\alias{str_datestamp}
\title{str_datestamp}
\usage{
str_datestamp(format = "\%Y-\%m-\%d", date = Sys.Date(), ...)
}
\arguments{
\item{format}{defaults to \code{YYYY-mm-dd} (ISO 9601).}

\item{date}{defaults to \code{\link[=Sys.Date]{Sys.Date()}}.}

\item{...}{passed to \code{format()}.}
}
\description{
Output a datestamp.
}
