% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_datetime.R
\name{str_datetime}
\alias{str_datetime}
\title{str_datetime}
\usage{
str_datetime(format = "\%Y\%m\%d-\%H\%M\%S")
}
\arguments{
\item{format}{(character) see \code{\link[=strftime]{strftime()}}}
}
\value{
(character)
}
\description{
Returns the current date and time, as a formatted string.
}
