% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_hash.R
\name{str_hash}
\alias{str_hash}
\alias{str_crc32}
\alias{str_xxhash32}
\alias{str_xxhash64}
\alias{str_md5}
\title{str_hash}
\usage{
str_hash(
  ...,
  algo = c("md5", "sha1", "crc32", "sha256", "sha512", "xxhash32", "xxhash64",
    "murmur32", "spookyhash", "blake3"),
  serialize = FALSE
)

str_crc32(...)

str_xxhash32(...)

str_xxhash64(...)

str_md5(...)
}
\arguments{
\item{...}{one or more character vectors (must be same length)}
}
\description{
Hash one or more characters, element-wise and in parallel.
}
