% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_humanize.R
\name{str_humanize}
\alias{str_humanize}
\alias{str_humanize.integer}
\alias{str_humanize.numeric}
\alias{str_humanize.tbl_df}
\title{Improve readability for humans}
\usage{
str_humanize(x, ...)

\method{str_humanize}{integer}(x, ...)

\method{str_humanize}{numeric}(x, digits = NULL, tiny = NULL, zero = NULL, ...)

\method{str_humanize}{tbl_df}(x, ...)
}
\arguments{
\item{x}{an integer or numeric vector, or a \code{\link[dplyr]{tbl}}}

\item{\dots}{(unused)}

\item{digits}{number of digits to retain}

\item{tiny}{what to print in place of tiny values (that round to zero, given the number of digits you've specified)}

\item{zero}{what to print in place of actual zeros (that are EXACTLY zero)}
}
\description{
Improve readability for humans
}
\section{Methods (by class)}{
\itemize{
\item \code{integer}: Do nothing to integers

\item \code{numeric}: Round real numbers to a given number of digits

\item \code{tbl_df}: Apply \code{str_humanize(...)} to each column in turn
}}

\note{
Per the help for \link{round}, the IEC 60559 ("round to the even digit") is used.
See \code{help(round)} (look under the Details section).
}
\examples{
x <- c(123456789.01, 0.996, 0.003, 0.007, 0.000001, 0)
str_humanize(x)
str_humanize(x, digits = 3)
str_humanize(x, digits = 3, tiny = "-")
str_humanize(x, digits = 2, tiny = "-", zero = "")

}
