% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_NOx.R
\name{str_NOx}
\alias{str_NOx}
\title{str_NOx}
\usage{
str_NOx(
  format = c("character", "utf8", "html", "TeX", "markdown"),
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{format}{(character) one of "character", "utf8", "html", "TeX", or "markdown"}

\item{verbose}{(logical)}
}
\value{
representation of "NOx" (nitrogen oxides)
}
\description{
str_NOx
}
