% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_or.R
\name{str_or}
\alias{str_or}
\title{str_or}
\usage{
str_or(x, ...)
}
\description{
A more human-friendly version of \code{str_csv()}.
}
\examples{
x <- c("apples", "oranges", "pears")
str_or(x)

}
\seealso{
\code{\link[=str_and]{str_and()}}
}
