% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_permute.R
\name{str_permute}
\alias{str_permute}
\title{Extract and collapse substrings}
\usage{
str_permute(x, ...)
}
\description{
Extract and collapse substrings
}
\examples{
ALPHABET <- paste0(LETTERS, collapse = "")
str_permute(rep(ALPHABET, 3), 5:6, 3:4)

}
