% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_PM25.R
\name{str_PM25}
\alias{str_PM25}
\title{str_PM25}
\usage{
str_PM25(
  format = c("character", "utf8", "html", "TeX", "markdown"),
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{format}{choice of output format}

\item{verbose}{(logical)}
}
\description{
str_PM25
}
