% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_prefix.R
\name{str_prefix}
\alias{str_prefix}
\title{str_prefix}
\usage{
str_prefix(x, prefix, list = seq_along(x))
}
\arguments{
\item{x}{(character) to be prefixed}

\item{list}{(optional) indices of elements to prefix; defaults to all}

\item{suffix}{(character) length one, or will be recycled}
}
\value{
(character)
}
\description{
str_prefix
}
\examples{
str_prefix(1:3, "S-")
str_prefix(1:5, "<", 1)

}
\seealso{
\itemize{
\item \code{\link[=str_suffix]{str_suffix()}}
}
}
