% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_printable.R
\name{str_printable}
\alias{str_printable}
\title{str_printable}
\usage{
str_printable(x, non_printable = "[^[:print:]]")
}
\description{
str_printable
}
\examples{
x <- as.raw(c(0x28, 0x12, 0x41))
print(rawToChar(x))
str_printable(rawToChar(x))

}
