% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_quote.R
\name{str_quote}
\alias{str_quote}
\title{Surround strings with quotation marks}
\usage{
str_quote(x, ch = "\\"")
}
\arguments{
\item{x}{character vector}

\item{ch}{character representing a quotation mark (default is double quote)}
}
\description{
Surround strings with quotation marks
}
\examples{
str_quote(LETTERS)

}
