% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_relevel.R
\name{str_relevel}
\alias{str_relevel}
\title{Relevel a character vector}
\usage{
str_relevel(x, what)
}
\arguments{
\item{x}{(character) vector}

\item{what}{(character) value to look for}
}
\description{
Move the first occurrence of a value to the first position.
}
\examples{
str_relevel(LETTERS, "J")
str_relevel(LETTERS, "foo")

}
