% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_sign.R
\name{str_sign}
\alias{str_sign}
\title{str_sign}
\usage{
str_sign(x)
}
\arguments{
\item{x}{numeric}
}
\value{
one of "+", "-", or "".
}
\description{
Return the signs of signed numbers.
Zero, NA, and NaN yield empty strings.
}
