% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_suffix.R
\name{str_suffix}
\alias{str_suffix}
\title{str_suffix}
\usage{
str_suffix(x, suffix, list = seq_along(x))
}
\arguments{
\item{x}{(character) to be suffixed}

\item{suffix}{(character) length one, or will be recycled}

\item{list}{(optional) indices of elements to suffix; defaults to all}
}
\value{
(character)
}
\description{
str_suffix
}
\examples{
str_suffix(1:2, "m")
str_suffix(1:5, " or more", 3)
str_suffix(1:3, c("st", "nd", "rd"))

}
\seealso{
\itemize{
\item \code{\link[=str_prefix]{str_prefix()}}
}
}
