% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_tex.R
\name{str_tex}
\alias{str_tex}
\title{str_tex}
\usage{
str_tex(...)
}
\arguments{
\item{...}{(character) combined via \link[stringr:str_glue(...)]{stringr::str_glue(...)}.}
}
\value{

}
\description{
Glues its \code{...} argument(s) together, then runs the result through \code{\link[latex2exp:TeX]{latex2exp::TeX()}}.
}
