% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_ugm3.R
\name{str_ugm3}
\alias{str_ugm3}
\title{str_ugm3}
\usage{
str_ugm3(
  format = c("character", "utf8", "html", "TeX"),
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{format}{(character) one of "character", "utf8", "html", or "TeX"}

\item{verbose}{(logical)}
}
\value{
representation of "ug/m3" (micrograms per cubic meter)
}
\description{
str_ugm3
}
