% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_integers.R
\name{unpack_integers}
\alias{unpack_integers}
\title{Unpack a vector of "packed integers" (like facility IDs or category IDs)}
\usage{
unpack_integers(input_data, var_name, ...,
  verbose = getOption("verbose"))
}
\description{
Unpack a vector of "packed integers" (like facility IDs or category IDs)
}
\examples{
df <- data_frame(cat_ids = c("c(1, 3)", "c(1:4)"))
unpack_integers(df, var_name = "cat_ids")
unpack_integers(df)

}
