% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_digits.R
\name{format_digits}
\alias{format_digits}
\title{Format with a fixed number of digits, keeping trailing zeros}
\usage{
format_digits(x, digits, ...)
}
\arguments{
\item{x}{numeric}

\item{digits}{number of digits to keep}

\item{...}{passed to \code{\link[=formatC]{formatC()}}}
}
\description{
Shorthand for \code{formatC(x, format="f", flag="#", digits=digits, ...)}.
}
