% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_SI.R
\name{format_SI}
\alias{format_SI}
\title{Format according to SI conventions}
\usage{
format_SI(x, ...)
}
\arguments{
\item{x}{(numeric)}

\item{...}{further arguments to \link[base:format]{format}}
}
\description{
Format according to SI conventions
}
\examples{
format_SI(1)
format_SI(9:11)
format_SI(c(33, 333, 3333, 33333333))

}
\seealso{
(Stack Overflow post)\link{http://stackoverflow.com/questions/21045545/how-to-accurately-display-si-prefix-for-numbers-in-y-axis-scale-of-plot-made-wit}
}
