% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_digits.R
\name{guess_digits}
\alias{guess_digits}
\title{guess_digits}
\usage{
guess_digits(x, digits = 0)
}
\description{
Really crude heuristic (!) and linear search (!) for an approximate precision.
}
\examples{
guess_digits(1)
guess_digits(0.66)
guess_digits(0.667)
}
\seealso{
\code{\link[=format_decimal]{format_decimal()}}
}
