% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_find.R
\name{str_find}
\alias{str_find}
\title{Identify rows containing key words}
\usage{
str_find(df, x, k.word, case.sensitive = TRUE)
}
\arguments{
\item{df}{A data frame containing the variable to be analyzed.}

\item{x}{Column name (character) or column index (numeric) indicating the variable to evaluate.}

\item{k.word}{A string vector that contains keyword(s) to be detected from the data frame.}

\item{case.sensitive=TRUE}{A logical evaluating to TRUE or FALSE indicating whether the function evaluates for uppercase or lowercase strings.}
}
\description{
Extract and report all instances of keywords defined in a string-based vector x.
}
\examples{
p <- c("jackal","panther")
str_find(animals_df, 2, p, case.sensitive=T)

}
