% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_find.R
\name{str_find}
\alias{str_find}
\title{Finding rows containing specified keywords}
\usage{
str_find(df, x, k.word, case.sensitive = TRUE)
}
\arguments{
\item{df}{a data frame containing the variable to be analyzed.}

\item{x}{column name (character) or column index (numeric) indicating the variable to evaluate.}

\item{k.word}{a string vector that contains keyword(s) to be detected from the data frame.}

\item{case.sensitive}{a logical evaluating to \code{TRUE} or \code{FALSE} indicating whether the function evaluates for uppercase or lowercase strings.}
}
\value{
A data frame containing only the rows where the selected column matches at least one of the specified keywords.
A message is printed indicating the total number of matches found.

If no matches are found, an empty data frame is returned with a message indicating no matching values were found.
}
\description{
This function searches for specified keyword(s) within a selected data frame column and returns all rows where a match is found.
It is useful for filtering observations based on string values.
}
\details{
If \code{case.sensitive = TRUE}, matching requires exact agreement in letter casing (i.e \code{"PANTHER"} and \code{"Panther"} will be treated as different values).

If \code{case.sensitive = FALSE}, matching ignores differences in letter casing (i.e, \code{"PANTHER"} and \code{"Panther"} are treated as equal).
}
\examples{


t <- c("eagles", "beetle", "falcon")
str_find(animals.df,"species", t, TRUE)

#Case.sensitive = TRUE
str_find(letters.df, 2,'Y', case.sensitive = TRUE)

#Case.sensitive = FALSE
str_find(letters.df, 2, 'Y', case.sensitive = FALSE)

}
