% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_unique.R
\name{str_unique}
\alias{str_unique}
\title{Summarize unique string values}
\usage{
str_unique(
  x,
  case.sensitive = TRUE,
  top_n = 10,
  plot = TRUE,
  use_percentage = FALSE,
  ...
)
}
\arguments{
\item{x}{a vector containing character or character-coercible values.}

\item{case.sensitive}{a logical value. If \code{TRUE}, unique values consider for letter casing (i.e \code{"TREE"} and \code{"tree"} will be treated as different values). If \code{FALSE}, all string values are set to lowercase (i.e, \code{"PANTHER"} is set to \code{"panther"}). Default is \code{TRUE}.}

\item{top_n}{a numeric value indicating how many of the most frequent values the bar plot shows. Default is 10.}

\item{plot}{a logical value. If \code{TRUE}, a bar plot showing the top 10 frequent string values will be shown. If \code{FALSE}, no bar plot will be created.}

\item{use_percentage}{a logical value. If \code{FALSE}, the y-axis and y-labels will use the \code{Count} data. If \code{TRUE}, the y-axis and y-labels will use the \code{Percentage} data.}

\item{...}{further arguments passed to customize bar plot visualization.}
}
\value{
A descending sorted matrix where each row corresponds to a unique value from the string vector. The matrix contains two columns:
\itemize{
\item \code{Count}: Number of occurrences of each unique value.
\item \code{Percentage}: Proportion of each value in the dataset.
}
If \code{plot} is set to \code{TRUE}, a bar plot will be created using the sorted matrix data.

If \code{use_percentage} is \code{TRUE}, the bar plot will use Percentage data rather than the default \code{Count} data.
}
\description{
\code{str_unique()} identifies all unique string values in a selected vector and returns a sorted summary of the count and proportion of each unique string.
A bar plot can also be produced showing the top n frequent values based on count or percentage.
}
\details{
The effectiveness of this function is reliant on how noisy the string data is.
}
\examples{

#Example with use_percentage = TRUE and top_n = 5
str_unique(letters.df$mixed_random.char, case.sensitive = TRUE, top_n = 5, plot = TRUE, use_percentage = TRUE)

#Example with use_percentage = FALSE, case.sensitive = FALSE and color = red
str_unique(letters.df$mixed_random.char, case.sensitive = FALSE, top_n = 10, plot=TRUE, use_percentage = FALSE, col = 'red')

#Example with character-coerced numeric columns
str_unique(mtcars$cyl)
}
